Renamed the package com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util to just com.brentdunklau.telepatriot_android.util.  This package got the wrong name a long time ago.  Finally fixing it.
