+++ app/src/main/java/com/brentdunklau/telepatriot_android/ChatFragment.java
@@ -12,19 +12,16 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
 
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.ChatMessage;
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.ChatMessageHolder;
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.User;
+import com.brentdunklau.telepatriot_android.util.ChatMessage;
+import com.brentdunklau.telepatriot_android.util.ChatMessageHolder;
+import com.brentdunklau.telepatriot_android.util.User;
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-
-import org.w3c.dom.Text;
 
 /**
  * Created by bdunklau on 10/11/17.
