+++ app/src/main/java/com/brentdunklau/telepatriot_android/LimboActivity.java
@@ -2,23 +2,18 @@ package com.brentdunklau.telepatriot_android;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.AccountStatusEvent;
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.AccountStatusEventHolder;
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.OneTime;
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.RoleAssignedListener;
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.User;
+import com.brentdunklau.telepatriot_android.util.AccountStatusEvent;
+import com.brentdunklau.telepatriot_android.util.AccountStatusEventHolder;
+import com.brentdunklau.telepatriot_android.util.OneTime;
+import com.brentdunklau.telepatriot_android.util.RoleAssignedListener;
+import com.brentdunklau.telepatriot_android.util.User;
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.ChildEventListener;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.FirebaseDatabase;
 
 /**
