+++ app/src/main/java/com/brentdunklau/telepatriot_android/ListUsersFragment.java
@@ -8,8 +8,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util.PassInfo;
-import com.google.firebase.database.FirebaseDatabase;
+import com.brentdunklau.telepatriot_android.util.PassInfo;
 
 import java.util.Map;
