+++ app/src/main/java/com/brentdunklau/telepatriot_android/util/AccountStatusEvent.java
@@ -1,4 +1,4 @@
-package com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util;
+package com.brentdunklau.telepatriot_android.util;
 
 /**
  * Created by bdunklau on 10/2/2017.
