+++ app/src/main/java/com/brentdunklau/telepatriot_android/util/ChatMessage.java
@@ -1,4 +1,4 @@
-package com.brentdunklau.telepatriot_android.com.brentdunklau.telepatriot_android.util;
+package com.brentdunklau.telepatriot_android.util;
 
 /**
  * Created by bdunklau on 10/10/2017.
